fix conflicts 2

+++ c_file/ls3.c
@@ -1,7 +1,7 @@
 
 int a = 1000;
 int main(int c ,int b){
-a = (int)c*100 + a;
+a = c*100 + a;
 c = 01000;
 c = 'a';
 c = c * b + (a - 100) * b;
